refactor: move transition logic to separate store
