+++ frontend/src/stores/element.js
@@ -16,6 +16,7 @@ import { generateUniqueId, cloneObj } from '../utils/helpers'
 import { guessTextColorFromBackground } from '../utils/color'
 import { handleUploadedMedia } from '../utils/mediaUploads'
 import { isPublicPresentation, presentationId } from './presentation'
+import { getUpdatedIdAfterConnections } from './transition'
 
 import { generateHTML } from '@tiptap/core'
 import { extensions } from '@/stores/tiptapSetup'
@@ -259,7 +260,7 @@ const replaceMediaElement = async (element, fileDoc) => {
 	if (element.type == 'video') {
 		element.poster = await getVideoPoster(fileDoc.file_url)
 	}
-	element.id = getUpdatedIdForElementContent(element)
+	element.id = getUpdatedIdAfterConnections(element)
 }
 
 const getDuplicateElementId = (element, srcSlide) => {
@@ -473,72 +474,11 @@ const addFixedWidthToElement = (deltaWidth) => {
 const { initTextEditor, activeEditor } = useTextEditor()
 let editorOldText = ''
 
-const compareTextContentForHTML = (currentContent, nextContent) => {
-	const parser = new DOMParser()
-
-	const doc = parser.parseFromString(currentContent, 'text/html')
-	const nextDoc = parser.parseFromString(nextContent, 'text/html')
-
-	const currentBlocks = doc.body.children
-	const nextBlocks = nextDoc.body.children
-
-	if (currentBlocks.length != nextBlocks.length) return false
-
-	for (let i = 0; i < currentBlocks.length; i++) {
-		// within each block, compare text content only (ignore styles and tags)
-		if (currentBlocks[i].textContent != nextBlocks[i].textContent) {
-			return false
-		}
-	}
-
-	return true
-}
-
-const getReferenceElement = (nextElement, slide) => {
-	for (const element of slide.elements) {
-		if (element.type != nextElement.type) continue
-
-		const hasSameText =
-			element.type == 'text' &&
-			compareTextContentForHTML(element.content, nextElement.content)
-		const hasSameSrc = element.type != 'text' && element.src == nextElement.src
-
-		if (hasSameText || hasSameSrc) {
-			return element
-		}
-	}
-}
-
-const getUpdatedIdForElementContent = (element, currentText) => {
-	const prevSlide = slides.value[slideIndex.value - 1]
-	const nextSlide = slides.value[slideIndex.value + 1]
-
-	let candidateSlide = null
-	if (prevSlide?.transition === 'Magic Move') {
-		// if transition begins on previous slide -
-		// check if any element in current slide refers to element from previous slide
-		candidateSlide = prevSlide
-	} else if (currentSlide.value?.transition === 'Magic Move' && nextSlide) {
-		// if transition begins on current slide -
-		// check if any element in next slide refers to this element
-		candidateSlide = nextSlide
-	}
-
-	if (candidateSlide) {
-		// find reference element in candidate slide
-		const refElement = getReferenceElement(element, candidateSlide)
-		// if found copy its id so that it does not re-render during transition
-		if (refElement) return refElement.id
-	}
-
-	return generateUniqueId()
-}
-
 const updateElementContent = (element) => {
 	const currentText = activeEditor.value.getText()
 	if (editorOldText == currentText) return
 
-	element.id = getUpdatedIdForElementContent(element)
+	element.id = getUpdatedIdAfterConnections(element)
 	element.content = activeEditor.value.getHTML()
 	editorOldText = currentText
 }
@@ -642,5 +582,4 @@ export {
 	normalizeZIndices,
 	isWithinOverlappingBounds,
 	updatePosition,
-	getReferenceElement,
 }
